requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pymupdf==1.23.26
sentence-transformers==2.3.1
scikit-learn==1.4.0
streamlit==1.31.0
requests==2.31.0
google-generativeai==0.3.2
pydantic==2.5.3